整理方法分类
