--- internal/app/server/common/tts/tts.go
@@ -1,1 +1,0 @@
-package websocket
